fixes
• REQUIREMENT : Two people must not be able to book the same seat – If I
open two browsers (I used Chrome and another instance of chrome with
private browsing turned on), login to a different user on each, it is
possible to book the same seat for the same conference with both users.
• REQUIREMENT : A single person may book up to four seats in one
transaction – I found that I could select up to 5 seats, it was only
when I tried to book the 6th seat that the error message came up telling
me the limit was 4 seats.
• REQUIREMENT : A unique name and email address are required for each
seat that is booked – Perhaps this requirement isn’t clear, the
intention being that I might want to reserve four seats for an event,
one for myself and three others for friends, and that each of those
seats would need their own unique name and email address. I couldn’t see
where to enter the unique details for each of the seats.